Add files via upload
In this project we have done complete data analysis and cluster players according to their skills and performed some given tasks. For model implementation of this dataset, we have used K-means, XGboosting and Logistic Regressioin algorithms. Since we have stated that K means clustering is the best model for Fifa20.